server/requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pycryptodome==3.17.0
sqlalchemy==2.0.12
aiosqlite==0.19.0

# Email support
email-validator==2.0.0
jinja2==3.1.2
python-dotenv==1.0.0

# Kafka integration
aiokafka==0.8.1
kafka-python==2.0.2

# Advanced security
pysgx==0.4.5  # Python bindings for Intel SGX
cryptography==40.0.2  # Enhanced cryptographic functions
argon2-cffi==21.3.0  # Stronger password hashing

# Query optimization
numpy==1.24.3  # For numerical operations
pandas==2.0.1  # For data manipulation
dask==2023.5.0  # For parallel computing

# Monitoring and logging
prometheus-fastapi-instrumentator==6.0.0
python-json-logger==2.0.7
